+++ src/main/java/at/hannibal2/skyhanni/events/EntityMoveEvent.kt
@@ -1,5 +1,6 @@
 package at.hannibal2.skyhanni.events
 
+import at.hannibal2.skyhanni.utils.LorenzVec
 import net.minecraft.entity.Entity
 
-class EntityMoveEvent(val entity: Entity) : LorenzEvent()
+class EntityMoveEvent(val entity: Entity, val oldLocation: LorenzVec, val newLocation: LorenzVec) : LorenzEvent()